cute getLabel Data.Weight => float64

+++ src/Entry/LabelInfo.go
@@ -20,7 +20,7 @@ type LabelData struct {
 	TN2                   interface{} `json:"TN2"`
 	Carrier               string      `json:"Carrier"`
 	CarrierServiceLevel   string      `json:"CarrierServiceLevel"`
-	Weight                int         `json:"Weight"`
+	Weight                float64     `json:"Weight"`
 	ShippingPrice         float64     `json:"ShippingPrice"`
 	ChargeDetail          interface{} `json:"ChargeDetail"`
 }
